+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 *   (improvement) Clear the output directory when writing watched JS files.
 *   (bc) Changed the storage place of the `_dependencies.json` (it is now in the root of the `output` path).
 *   (bc) Changed the structure and keys in the dependencies map, to also allow SCSS files to be written into the dependencies map.
+*   (feature) Add option to disable the legacy build.
 
 
 1.2.12